--- src/spetlr/deltaspec/DeltaTableSpecDifference.py
@@ -1,1 +1,0 @@
-
