+++ src/spetlr/deltaspec/exceptions.py
@@ -1,4 +1,7 @@
-class InvalidSpecificationError(Exception):
+from spetlr.exceptions import SpetlrException
+
+
+class InvalidSpecificationError(SpetlrException):
     """Internally this exception is used when attempting
     to unpack configurations into table and db specifications and this fails"""
 
@@ -7,5 +10,11 @@ class InvalidSpecificationError(Exception):
 
 class NoTableAtTarget(InvalidSpecificationError):
     """Indiactes thate there is not table at the target location."""
+
+    pass
+
+
+class TableSpecNotReadable(SpetlrException):
+    """A read was attempted, but the table is not readable as specified."""
 
     pass
